third_party/blink/web_tests/external/wpt/navigation-api/focus-reset/autofocus_same-checkpoint-expected.txt

This is a testharness.js-based test.
[FAIL] An element with autofocus, present before navigation, gets focused
  assert_equals: Focus stays on the non-autofocused button after the transition expected Element node <button autofocus=""></button> but got Element node <button></button>
[FAIL] Two elements with autofocus, present before navigation; the first gets focused
  assert_equals: Focus stays on the non-autofocused button after the transition expected Element node <button autofocus=""></button> but got Element node <button autofocus=""></button>
[FAIL] An element with autofocus, present before navigation but disabled before finished, does not get focused
  assert_equals: Focus stays on the non-autofocused button after the transition expected Element node <button autofocus=""></button> but got Element node <button></button>
[FAIL] An element with autofocus, present before navigation but with its autofocus attribute removed before finished, does not get focused
  assert_equals: Focus stays on the non-autofocused button after the transition expected Element node <button autofocus=""></button> but got Element node <button></button>
[FAIL] Two elements with autofocus, present before navigation, but the first gets disabled; the second gets focused
  assert_equals: Focus stays on the initially-focused button during the transition expected Element node <button autofocus="" disabled=""></button> but got Element node <button autofocus=""></button>
[FAIL] An element with autofocus, introduced between committed and finished, gets focused
  assert_equals: Focus gets reset after the transition expected Element node <body>\n\n<script type="module">\n\nconst delayed = location.... but got Element node <button></button>
[FAIL] An element with autofocus, introduced after finished, does not get focused
  assert_equals: Focus gets reset after the transition expected Element node <body>\n\n<script type="module">\n\nconst delayed = location.... but got Element node <button></button>
Harness: the test ran to completion.
